% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{eq_crownwidth}
\alias{eq_crownwidth}
\title{The equation to get the crown width of tree}
\usage{
eq_crownwidth(dbh, a = 2, b = 0.24)
}
\arguments{
\item{b}{}
}
\description{
The equation to get the crown width of tree
}
