% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{eq_height}
\alias{eq_height}
\title{The allometric equation of DBH (D) and Height (H) of tree in the isolated area.}
\usage{
eq_height(dbh, alpha = 0.8, beta = 0.75)
}
\arguments{
\item{dbh}{The diameter at breast height (DBH) in cm}

\item{alpha}{The scaling exponent}

\item{beta}{The proportionality coefficient}
}
\value{
Height of tree in meter (m)
}
\description{
The equation is H = alpha*D^beta
The default parameters (alpha=0.8, beta=0.75) is for Hevea brasiliensis species
}
