% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfba.R
\name{fba.branches}
\alias{fba.branches}
\title{Functional Branches Analysis}
\usage{
fba.branches(
  initial_d = 10,
  showProgress = FALSE,
  Intercept = 21.944,
  Slope = 8.462,
  RangeL = 0.2,
  LocalityTC = 0.001,
  ScaleTC = 5e-04,
  Thickening = 0,
  DMin = 0.8,
  RangeDmin = 0.5,
  AvNsub = 2.687150838,
  HowP = 0,
  LocalityP = 1,
  ScaleP = 0.1,
  Prob1P = 0.0279329608938547,
  Prob2P = 0.134078212290503,
  Prob3P = 0.318435754189944,
  Prob4P = 0.240223463687151,
  Prob5P = 0.134078212290503,
  Prob6P = 0.145251396648045,
  Prob1Q = 0.206703910614525,
  Prob2Q = 0.201117318435754,
  Prob3Q = 0.162011173184358,
  Prob4Q = 0.122905027932961,
  Prob5Q = 0.150837988826816,
  BranchQ = NULL,
  InterceptTwig = 22.657,
  SlopeTwig = 1.4108,
  RangeLTwig = 0.3,
  InterceptBranch = 23,
  SlopeBranch = 0,
  RangeLBranch = 0.5,
  InterceptWood = 120,
  SlopeWood = 0,
  RangeLWood = 0.8,
  InterceptStem = 66,
  SlopeStem = 20,
  RangeLStem = 0.5,
  MaxDiamTwig = 2,
  MaxDiamBranch = 8,
  singleLD = 0,
  LBaretip = 0,
  Dmaxfin = 1.533333333,
  Dzerofin = 2.9,
  Maxfindens = 0.499987708,
  OrthoPlagio = 0,
  DLcontinuous = 1,
  Randseed = NULL,
  orthoVertAnglesMaxDeg = 5,
  plagioVertAnglesDeg = 85,
  plagioVertAnglesSDDeg = 5,
  twigVertAnglesDeg = 70,
  twigVertAnglesSDDeg = 20
)
}
\arguments{
\item{Randseed}{}
}
\description{
Compute the branches dimensions and return the list of branches segments
}
