% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate Tree object}
\usage{
generate_tree(
  dbh = 15,
  height = NULL,
  crown_base_height = NULL,
  crown_width_max_height = NULL,
  crown_radius = NULL,
  crown_radius_init_angle = 0
)
}
\arguments{
\item{dbh}{Tree size}

\item{crown_radius_init_angle}{}
}
\value{
Tree object
}
\description{
The function for generating Tree object.
The object can be viewed using vtree().
}
