% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{generate_triangle_idx}
\alias{generate_triangle_idx}
\title{generate_triangle_idx}
\usage{
generate_triangle_idx(
  lateral_n = LATERAL_DIMENSION,
  vertical_n = VERTICAL_DIMENSION
)
}
\arguments{
\item{vertical_n}{}
}
\description{
generate_triangle_idx
}
