% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{generate_vbranches}
\alias{generate_vbranches}
\title{Generate branches}
\usage{
generate_vbranches(
  crown_depth = 3,
  crown_curve_depth = 2,
  crown_radius = 1,
  crown_radius_init_angle = 0
)
}
\arguments{
\item{crown_radius_init_angle}{}
}
\description{
Generate branches
}
