% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeforest.R
\name{threeforest}
\alias{threeforest}
\title{The widget for displaying trees in 3D}
\usage{
threeforest(
  forest_obj,
  setting = list(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
This widget using THREEJS library for 3D visualization
}
