% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\name{vtree}
\alias{vtree}
\title{Display trees in 3D visualization}
\usage{
vtree(trees, setting = list(), width = NULL, height = NULL, elementId = NULL)
}
\description{
Trees object should be generated by generate_trees()
}
